requirements.txt
glob2
matplotlib
numpy==1.18.3
scikit-image
albumentations
pytorch_model_summary
tqdm
thop
#--find-links https://download.pytorch.org/whl/torch_stable.html
#torch==1.7.1+cu110
#torchvision==0.8.2+cu110

